feat: create multiple spawn configs per creature
